muokattu entityt ja repot uuden tietokantamallin mukaisiksi
