+++ Poistetut/AlennusRepository.java
@@ -1,7 +1,8 @@
-package ohjelmistoprojekti1.a3004.domain;
+/* package ohjelmistoprojekti1.a3004.domain;
 
 import org.springframework.data.repository.CrudRepository;
 
 public interface AlennusRepository extends CrudRepository<Alennus, Long>{
 
 }
+ */
