+++ Poistetut/TapahtumajarjestajaHenkiloRepository.java
@@ -1,0 +1,8 @@
+/* package ohjelmistoprojekti1.a3004.domain;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface TapahtumajarjestajaHenkiloRepository extends CrudRepository<TapahtumajarjestajaHenkilo, Long> {
+
+}
+ */
