+++ src/main/java/ohjelmistoprojekti1/a3004/domain/MyyntitapahtumaRepository.java
@@ -3,9 +3,5 @@ package ohjelmistoprojekti1.a3004.domain;
 import org.springframework.data.repository.CrudRepository;
 
 public interface MyyntitapahtumaRepository extends CrudRepository<Myyntitapahtuma, Long> {
-    // tässä luodaan normaali luokka, joka nimetään entity-luokka + Repository
-    // muut muutokset voi tehdä manuaalisesti, eli
-    // muuttaa: class -> interface
-    // ja lisätä: extends CrudRepository<Entity-luokka, Entity-luokan primary key:n tietotyyppi>
 
 }
