+++ src/main/java/ohjelmistoprojekti1/a3004/domain/Tapahtuma.java
@@ -1,13 +1,16 @@
 package ohjelmistoprojekti1.a3004.domain;
 
 import java.time.LocalDate;
+import java.util.List;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+
+import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 
 @Entity
@@ -20,43 +23,35 @@ public class Tapahtuma {
     private String tapahtuman_nimi;
     private String paikka, katuosoite;
     private LocalDate alku_pvm, loppu_pvm;
-    private int lippu_lukum, perus_hinta;
+    private int lippu_lukum;
 
-    /* NÄMÄ KOMMENTIT VOI AVATA KUN TAPAHTUMANJARJESTAJA TAULU ON VALMIS 
+    @OneToMany(cascade = CascadeType.ALL, mappedBy = "tapahtuman_lipputyyppi")
+    @JsonIgnore
+    private List<TapahtumanLipputyyppi> tapahtuman_lipputyypit;
 
-    @ManyToOne
+/*     @ManyToOne
     @JoinColumn(name = "jarjestaja_id")
-    private Tapahtumanjarjestaja tapahtumanjarjestaja;
-     */
+    private Tapahtumanjarjestaja tapahtumanjarjestaja; */
 
- /* NÄMÄ KOMMENTIT VOI AVATA KUN POSTINUMERO TAULU ON VALMIS 
-    @ManyToOne
+/*     @ManyToOne
     @JoinColumn(name = "postinumero_id")
-    private Postinumero postinumero;
-
-     */
+    private Postinumero postinumero; */
+    
     //Parametriton konstruktori
     public Tapahtuma() {
     }
 
-    /* NÄMÄ KOMMENTIT VOI AVATA KUN POSTINUMERO JA TAPAHTUMANJARJESTAJA TAULUT OVAT VALMIITA 
-
-    //Parametrillinen konstruktori, ei ID:tä, koska se autogeneroituu.
     public Tapahtuma(String tapahtuman_nimi, String paikka, String katuosoite, LocalDate alku_pvm, LocalDate loppu_pvm,
-            int lippu_lukum, int perus_hinta, Tapahtumanjarjestaja tapahtumanjarjestaja, Postinumero postinumero) {
+            int lippu_lukum, List<TapahtumanLipputyyppi> tapahtuman_lipputyypit) {
         this.tapahtuman_nimi = tapahtuman_nimi;
         this.paikka = paikka;
         this.katuosoite = katuosoite;
         this.alku_pvm = alku_pvm;
         this.loppu_pvm = loppu_pvm;
         this.lippu_lukum = lippu_lukum;
-        this.perus_hinta = perus_hinta;
-        this.tapahtumanjarjestaja = tapahtumanjarjestaja;
-        this.postinumero = postinumero;
+        this.tapahtuman_lipputyypit = tapahtuman_lipputyypit;
     }
 
-     */
-    //Getterit ja setterit
     public Long getTapahtuma_id() {
         return tapahtuma_id;
     }
@@ -113,45 +108,21 @@ public class Tapahtuma {
         this.lippu_lukum = lippu_lukum;
     }
 
-    public int getPerus_hinta() {
-        return perus_hinta;
+    public List<TapahtumanLipputyyppi> getTapahtuman_lipputyypit() {
+        return tapahtuman_lipputyypit;
     }
 
-    public void setPerus_hinta(int perus_hinta) {
-        this.perus_hinta = perus_hinta;
+    public void setTapahtuman_lipputyypit(List<TapahtumanLipputyyppi> tapahtuman_lipputyypit) {
+        this.tapahtuman_lipputyypit = tapahtuman_lipputyypit;
     }
 
-    /* NÄMÄ KOMMENTIT VOI AVATA KUN TAPAHTUMANJARJESTAJA TAULU ON VALMIS 
-
-    public Tapahtumanjarjestaja getTapahtumanjarjestaja() {
-        return tapahtumanjarjestaja;
-    }
-
-    public void setTapahtumanjarjestaja(Tapahtumanjarjestaja tapahtumanjarjestaja) {
-        this.tapahtumanjarjestaja = tapahtumanjarjestaja;
-    }
-
-     */
-
- /* NÄMÄ KOMMENTIT VOI AVATA KUN POSTINUMERO TAULU ON VALMIS 
-
-    public Postinumero getPostinumero() {
-        return postinumero;
-    }
-
-    public void setPostinumero(Postinumero postinumero) {
-        this.postinumero = postinumero;
-    }
-
-     */
-
- /* NÄMÄ KOMMENTIT VOI AVATA KUN POSTINUMERO JA TAPAHTUMANJARJESTAJA TAULUT OVAT VALMIITA 
     @Override
     public String toString() {
         return "Tapahtuma [tapahtuma_id=" + tapahtuma_id + ", tapahtuman_nimi=" + tapahtuman_nimi + ", paikka=" + paikka
                 + ", katuosoite=" + katuosoite + ", alku_pvm=" + alku_pvm + ", loppu_pvm=" + loppu_pvm
-                + ", lippu_lukum=" + lippu_lukum + ", perus_hinta=" + perus_hinta + ", tapahtumanjarjestaja="
-                + tapahtumanjarjestaja + ", postinumero=" + postinumero + "]";
+                + ", lippu_lukum=" + lippu_lukum + ", tapahtuman_lipputyypit=" + tapahtuman_lipputyypit + "]";
     }
-     */
+
+    
+    
 }
