--- src/main/java/ohjelmistoprojekti1/a3004/domain/TapahtumajarjestajaYritysRepository.java
@@ -1,7 +1,0 @@
-package ohjelmistoprojekti1.a3004.domain;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface TapahtumajarjestajaYritysRepository extends CrudRepository<jarjestaja_id, Long>{
-
-}
