+++ src/main/java/ohjelmistoprojekti1/a3004/domain/TapahtumanLipputyyppiRepository.java
@@ -1,0 +1,7 @@
+package ohjelmistoprojekti1.a3004.domain;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface TapahtumanLipputyyppiRepository extends CrudRepository<TapahtumanLipputyyppi, Long> {
+    
+}
