--- src/main/java/ohjelmistoprojekti1/a3004/domain/example.java
@@ -1,5 +1,0 @@
-package ohjelmistoprojekti1.a3004.domain;
-
-public class example {
-
-}
